+++ src/problem/constr/constr_ACPF.c
@@ -404,14 +404,12 @@ void CONSTR_ACPF_allocate(Constr* c) {
   int* H_nnz;
   int P_index;
   int Q_index;
-  Mat* H_array;
   Mat* HP;
   Mat* HQ;
   int* row;
   int* col;
   int i;
   int t;
-  int H_comb_nnz;
   int bus_index_t;
 
   net = CONSTR_get_network(c);
@@ -440,16 +438,14 @@ void CONSTR_ACPF_allocate(Constr* c) {
 			 J_nnz));  // nnz
 
   // H array
-  H_comb_nnz = 0;
-  H_array = MAT_array_new(num_constr);
-  CONSTR_set_H_array(c,H_array,num_constr);
+  CONSTR_allocate_H_array(c,num_constr);
   for (t = 0; t < num_periods; t++) {
     for (i = 0; i < num_buses; i++) {
       bus_index_t = i+t*num_buses;
       P_index = BUS_get_index_P(NET_get_bus(net,i))+t*2*num_buses;
       Q_index = BUS_get_index_Q(NET_get_bus(net,i))+t*2*num_buses;
-      HP = MAT_array_get(H_array,P_index);
-      HQ = MAT_array_get(H_array,Q_index);
+      HP = CONSTR_get_H_single(c,P_index);
+      HQ = CONSTR_get_H_single(c,Q_index);
       MAT_set_nnz(HP,H_nnz[bus_index_t]);
       MAT_set_nnz(HQ,H_nnz[bus_index_t]);
       MAT_set_size1(HP,num_vars);
@@ -471,15 +467,8 @@ void CONSTR_ACPF_allocate(Constr* c) {
 
       MAT_set_data_array(HP,(REAL*)malloc(H_nnz[bus_index_t]*sizeof(REAL))); // different data array
       MAT_set_data_array(HQ,(REAL*)malloc(H_nnz[bus_index_t]*sizeof(REAL)));
-
-      H_comb_nnz += 2*H_nnz[bus_index_t];
     }
   }
-
-  // H combined
-  CONSTR_set_H_combined(c,MAT_new(num_vars,     // size1 (rows)
-				  num_vars,     // size2 (cols)
-				  H_comb_nnz)); // nnz
 }
 
 void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
@@ -495,14 +484,9 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
   int* J_nnz;
   int* H_nnz;
   int H_nnz_val;
-  int H_nnz_comb;
   char* bus_counted;
   Mat* H_array;
   Mat* H[2];
-  int* Hi;
-  int* Hj;
-  int* Hi_comb;
-  int* Hj_comb;
   int bus_index_t[2];
   int w_index[2];
   int v_index[2];
@@ -516,15 +500,10 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
   int phi_index;
   int k;
   int m;
-  int temp;
   int num_buses;
-  int T;
 
   // Num buses
   num_buses = NET_get_num_buses(CONSTR_get_network(c));
-
-  // Num periods
-  T = BRANCH_get_num_periods(br);
 
   // Constr data
   J = CONSTR_get_J(c);
@@ -534,7 +513,7 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
   bus_counted = CONSTR_get_bus_counted(c);
 
   // Check pointers
-  if (!J_nnz || !H_nnz || !bus_counted)
+  if (!J_nnz || !H_nnz || !H_array || !bus_counted)
     return;
 
   // Check outage
@@ -903,29 +882,6 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
 
     // Update counted flag
     bus_counted[bus_index_t[k]] = TRUE;
-  }
-
-  // Done
-  if ((t == T-1) && (BRANCH_get_index(br) == NET_get_num_branches(CONSTR_get_network(c))-1)) {
-
-    // Ensure lower triangular and save struct of H comb
-    H_nnz_comb = 0;
-    Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
-    Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
-    for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
-      Hi = MAT_get_row_array(MAT_array_get(H_array,k));
-      Hj = MAT_get_col_array(MAT_array_get(H_array,k));
-      for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
-	if (Hi[m] < Hj[m]) {
-	  temp = Hi[m];
-	  Hi[m] = Hj[m];
-	  Hj[m] = temp;
-	}
-	Hi_comb[H_nnz_comb] = Hi[m];
-	Hj_comb[H_nnz_comb] = Hj[m];
-	H_nnz_comb++;
-      }
-    }
   }
 }