+++ src/problem/constr/constr_REG_GEN.c
@@ -184,9 +184,7 @@ void CONSTR_REG_GEN_allocate(Constr* c) {
   int A_row;
   int J_row;
   int* H_nnz;
-  Mat* H_array;
   Mat* H;
-  int H_comb_nnz;
   int num_vars;
   int num_extra_vars;
   int i;
@@ -199,9 +197,10 @@ void CONSTR_REG_GEN_allocate(Constr* c) {
   num_vars = NET_get_num_vars(CONSTR_get_network(c));
   num_extra_vars = CONSTR_get_num_extra_vars(c);
 
-  // Extra var limits
+  // Extra vars
   CONSTR_set_l_extra_vars(c,VEC_new(num_extra_vars));
   CONSTR_set_u_extra_vars(c,VEC_new(num_extra_vars));
+  CONSTR_set_init_extra_vars(c,VEC_new(num_extra_vars));
 
   // G u l
   CONSTR_set_G(c,MAT_new(0,num_vars+num_extra_vars,0));
@@ -225,24 +224,16 @@ void CONSTR_REG_GEN_allocate(Constr* c) {
 			 J_nnz));                 // nnz
 
   // H
-  H_comb_nnz = 0;
-  H_array = MAT_array_new(J_row);
-  CONSTR_set_H_array(c,H_array,J_row);
+  CONSTR_allocate_H_array(c,J_row);
   for (i = 0; i < J_row; i++) {
-    H = MAT_array_get(H_array,i);
+    H = CONSTR_get_H_single(c,i);
     MAT_set_nnz(H,H_nnz[i]);
     MAT_set_size1(H,num_vars+num_extra_vars);
     MAT_set_size2(H,num_vars+num_extra_vars);
     MAT_set_row_array(H,(int*)calloc(H_nnz[i],sizeof(int)));
     MAT_set_col_array(H,(int*)calloc(H_nnz[i],sizeof(int)));
     MAT_set_data_array(H,(REAL*)calloc(H_nnz[i],sizeof(REAL)));
-    H_comb_nnz += H_nnz[i];
   }
-
-  // H combined
-  CONSTR_set_H_combined(c,MAT_new(num_vars+num_extra_vars, // size1 (rows)
-				  num_vars+num_extra_vars, // size2 (cols)
-				  H_comb_nnz));            // nnz
 }
 
 void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
@@ -258,23 +249,16 @@ void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
   Mat* H_array;
   Mat* Hy;
   Mat* Hz;
-  int* Hi;
-  int* Hj;
-  int* Hi_comb;
-  int* Hj_comb;
   int* A_nnz;
   int* J_nnz;
   int* A_row;
   int* J_row;
   int* H_nnz;
-  int H_nnz_comb;
   char* bus_counted;
   int bus_index_t[2];
   int index_y;
   int index_z;
   int k;
-  int m;
-  int temp;
   int T;
   int num_vars;
 
@@ -295,7 +279,7 @@ void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
   bus_counted = CONSTR_get_bus_counted(c);
 
   // Check pointers
-  if (!A_nnz || !J_nnz || !A_row || !J_row || !H_nnz || !bus_counted)
+  if (!A_nnz || !J_nnz || !A_row || !H_array || !J_row || !H_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -442,29 +426,6 @@ void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
     // Update counted flag
     bus_counted[bus_index_t[k]] = TRUE;
   }
-
-  // Done
-  if ((t == T-1) && (BRANCH_get_index(br) == NET_get_num_branches(CONSTR_get_network(c))-1)) {
-
-    // Ensure lower triangular and save struct of H comb
-    H_nnz_comb = 0;
-    Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
-    Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
-    for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
-      Hi = MAT_get_row_array(MAT_array_get(H_array,k));
-      Hj = MAT_get_col_array(MAT_array_get(H_array,k));
-      for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
-	if (Hi[m] < Hj[m]) {
-	  temp = Hi[m];
-	  Hi[m] = Hj[m];
-	  Hj[m] = temp;
-	}
-	Hi_comb[H_nnz_comb] = Hi[m];
-	Hj_comb[H_nnz_comb] = Hj[m];
-	H_nnz_comb++;
-      }
-    }
-  }
 }
 
 void CONSTR_REG_GEN_eval_step(Constr* c, Branch* br, int t, Vec* values, Vec* values_extra) {
@@ -485,8 +446,6 @@ void CONSTR_REG_GEN_eval_step(Constr* c, Branch* br, int t, Vec* values, Vec* va
   char* bus_counted;
   int bus_index_t[2];
   int k;
-  REAL v;
-  REAL v_set;
   REAL y;
   REAL z;
   REAL Qsum;