+++ CHANGELOG.md
@@ -10,6 +10,12 @@ Targets
 
 Unreleased
 ----------
+* Support for init values for constraint extra variables (done).
+* Improved post processing of structures of constraint Hessians (now constr.c ensures lower triangular and fills H_combined ij, fixing bug in AC_FLOW_LIM) (done).
+* Mat parser detection of branches out of service (done).
+* H_combined is now completely handled by "base" constraint in constr.c. Custom constraints no longer need to allocate this matrix (done).
+* -Wall -Werror had no effect in Makefile.am and were moved to configure.ac. Now they work (requires autoconf-archive) (done).
+* Support for adding nonlinear constraints in Python and documentation (done).
 
 Version 1.3.0
 -------------